【20250326】證書題2-2

+++ AtCoder-374.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main(){
     double A, B;
     cin >> A >> B;
-    double S = static_cast<double> (B/A);
+    double S = B/A;
     cout << fixed << setprecision(3);
     cout << S << endl;
 }
